Vi/Vim fixes
After Homebrew reinstall of vim with the system override flag, I no
longer need to point mvim to the current version.

Remove the alias of vim pointing to mvim so that the system can use the
correct vim when a program calls for the system editor (such as Git
during commits).